@@ -1,6 +1,4 @@
-class Video < ActiveRecord::Base
-  include Publicable
-
+class Video < ApplicationRecord
   belongs_to :comedy
   has_one :youtube_link, as: :linker
 
