@@ -6,6 +6,6 @@ class ApplicationController < ActionController::Base
   private
 
   def load_random_comedies
-    @random_comedies = Comedy.publicable.random
+    @random_comedies = Comedy.random
   end
 end
