@@ -1,5 +1,5 @@
 class ImagesController < ApplicationController
   def index
-    @images = Image.comedy_poster.publicable.page(params[:page]).per(10)
+    @images = Image.comedy_poster.page(params[:page]).per(10)
   end
 end
