@@ -1,5 +1,5 @@
 class VideosController < ApplicationController
   def index
-    @videos = Video.publicable.page(params[:page]).per(5)
+    @videos = Video.page(params[:page]).per(5)
   end
 end
