@@ -11,32 +11,20 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20141028135422) do
+ActiveRecord::Schema.define(version: 20160403185327) do
 
-  # These are extensions that must be enabled in order to support this database
-  enable_extension "plpgsql"
-
-  create_table "categories", force: true do |t|
-    t.string   "title"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-  end
-
-  create_table "comedies", force: true do |t|
+  create_table "comedies", force: :cascade do |t|
     t.string   "cz_title"
     t.text     "text"
-    t.boolean  "publish"
-    t.integer  "category_id"
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "en_title"
     t.string   "slug"
-    t.datetime "published_at"
   end
 
-  add_index "comedies", ["slug"], name: "index_comedies_on_slug", unique: true, using: :btree
+  add_index "comedies", ["slug"], name: "index_comedies_on_slug", unique: true
 
-  create_table "images", force: true do |t|
+  create_table "images", force: :cascade do |t|
     t.string   "title"
     t.integer  "comedy_id"
     t.datetime "created_at"
@@ -46,11 +34,9 @@ ActiveRecord::Schema.define(version: 20141028135422) do
     t.integer  "image_file_size"
     t.datetime "image_updated_at"
     t.string   "alt"
-    t.boolean  "publish"
-    t.datetime "published_at"
   end
 
-  create_table "links", force: true do |t|
+  create_table "links", force: :cascade do |t|
     t.datetime "created_at"
     t.datetime "updated_at"
     t.string   "type"
@@ -59,41 +45,11 @@ ActiveRecord::Schema.define(version: 20141028135422) do
     t.string   "linker_type"
   end
 
-  create_table "posts", force: true do |t|
-    t.text     "text"
-    t.string   "title"
-    t.boolean  "publish"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-    t.text     "preface"
-    t.datetime "published_at"
-  end
-
-  create_table "users", force: true do |t|
-    t.string   "email",                  default: "", null: false
-    t.string   "encrypted_password",     default: "", null: false
-    t.string   "reset_password_token"
-    t.datetime "reset_password_sent_at"
-    t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          default: 0,  null: false
-    t.datetime "current_sign_in_at"
-    t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-  end
-
-  add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
-  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
-
-  create_table "videos", force: true do |t|
+  create_table "videos", force: :cascade do |t|
     t.string   "title"
     t.integer  "comedy_id"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "publish"
-    t.datetime "published_at"
   end
 
 end
